simplify Request.ts
