+++ lib/core/JSPError.ts
@@ -1,3 +1,6 @@
+/**
+ * @internal
+ */
 export default class JSPError extends Error {
   constructor (type: TError, description: string, err?: string) {
     super(description + (err ? '\n' + err : '')) // eslint-disable-line @typescript-eslint/strict-boolean-expressions
@@ -5,4 +8,7 @@ export default class JSPError extends Error {
   }
 }
 
+/**
+ * @internal
+ */
 type TError = 'InternalError' | 'APIError'
