+++ lib/static/api/v1.ts
@@ -1,3 +1,6 @@
+/**
+ * @internal
+ */
 export const api = {
   url: 'jspaste.tnfangel.xyz/',
   route: {
