+++ lib/static/messages.ts
@@ -1,3 +1,6 @@
+/**
+ * @internal
+ */
 export const error = {
   PACKAGE_UNDICI_MISSING: 'The use of the "undici" package is required, install "undici" or run a Node release greater than or equal to 18.0.0.'
 }
